[Backend][#69]Trying to fix Sockets
